Refactor: Relocate Invoice Generation Code and Enhance PDF Output
This commit reorganizes the codebase by moving invoice generation related files to a more generic `ui.invoice` package. It also enhances the generated PDF invoice by adding a company logo, an acknowledgment message, and a signature image.

**Key Changes:**

-   **File Relocation:**
    -   Files related to invoice canvas drawing (`CanvasAxis.kt`, `CanvasSentence.kt`, `CanvasHeadingView.kt`, `CanvasPdfView.kt`, `CanvasImage.kt`) have been moved from `net.techandgraphics.wastical.ui.screen.client.invoice.canvas` to `net.techandgraphics.wastical.ui.invoice.canvas`.
    -   Files related to PDF generation logic (`InvoiceToPdf.kt`, `PdfHeadingView.kt`, `PdfBgSentence.kt`, `PdfLineView.kt`, `PdfSentence.kt`, `PdfRectangle.kt`) have been moved from `net.techandgraphics.wastical.ui.screen.client.invoice.pdf` to `net.techandgraphics.wastical.ui.invoice.pdf`.

-   **PDF Enhancements (`InvoiceToPdf.kt`):**
    -   **Company Logo:** The company logo (`R.drawable.ic_logo`) is now drawn at the top-right corner of the PDF.
    -   **Acknowledgment & Signature:**
        -   An acknowledgment message "Received with thanks" in orange, bold text is added towards the bottom of the invoice.
        -   A signature image (`R.drawable.im_signature`) is placed below the acknowledgment message.

-   **ViewModel Updates:**
    -   `ClientInvoiceViewModel.kt`, `ClientHomeViewModel.kt`, `CompanyPaymentInvoiceViewModel.kt`, and `CompanyPaymentHistoryViewModel.kt` have been updated to import `invoiceToPdf` from the new `net.techandgraphics.wastical.ui.invoice.pdf` package.
    -   `PdfHeadingView.kt` and related files under `ui.invoice.pdf` updated to import `CanvasAxis` from the new `net.techandgraphics.wastical.ui.invoice.canvas` package.

**Impact:**
This refactoring improves the project structure by grouping common invoice functionality. The additions to the PDF provide a more professional and complete invoice document.
