+++ app/src/main/java/net/techandgraphics/wastical/ui/invoice/canvas/CanvasAxis.kt
@@ -1,0 +1,3 @@
+package net.techandgraphics.wastical.ui.invoice.canvas
+
+data class CanvasAxis(val xAxis: Float, val yAxis: Float)
