+++ app/src/main/java/net/techandgraphics/wastical/ui/invoice/canvas/CanvasSentence.kt
@@ -1,4 +1,4 @@
-package net.techandgraphics.wastical.ui.screen.client.invoice.canvas
+package net.techandgraphics.wastical.ui.invoice.canvas
 
 import android.graphics.Paint
 import androidx.compose.ui.graphics.drawscope.DrawScope
