+++ app/src/main/java/net/techandgraphics/wastical/ui/invoice/pdf/PdfHeadingView.kt
@@ -1,8 +1,8 @@
-package net.techandgraphics.wastical.ui.screen.client.invoice.pdf
+package net.techandgraphics.wastical.ui.invoice.pdf
 
 import android.graphics.Canvas
 import android.graphics.Paint
-import net.techandgraphics.wastical.ui.screen.client.invoice.canvas.CanvasAxis
+import net.techandgraphics.wastical.ui.invoice.canvas.CanvasAxis
 
 fun Canvas.pdfHeadingView(
   theHeading: String,
