+++ app/src/main/java/net/techandgraphics/wastical/ui/invoice/pdf/PdfRectangle.kt
@@ -1,4 +1,4 @@
-package net.techandgraphics.wastical.ui.screen.client.invoice.pdf
+package net.techandgraphics.wastical.ui.invoice.pdf
 
 import android.graphics.Canvas
 import android.graphics.Color
