--- app/src/main/java/net/techandgraphics/wastical/ui/screen/client/invoice/canvas/CanvasAxis.kt
@@ -1,3 +1,0 @@
-package net.techandgraphics.wastical.ui.screen.client.invoice.canvas
-
-data class CanvasAxis(val xAxis: Float, val yAxis: Float)
