+++ app/src/main/java/net/techandgraphics/wastical/ui/screen/company/client/invoice/CompanyPaymentInvoiceViewModel.kt
@@ -24,7 +24,7 @@ import net.techandgraphics.wastical.domain.toPaymentPlanUiModel
 import net.techandgraphics.wastical.domain.toPaymentWithMonthsCoveredUiModel
 import net.techandgraphics.wastical.preview
 import net.techandgraphics.wastical.share
-import net.techandgraphics.wastical.ui.screen.client.invoice.pdf.invoiceToPdf
+import net.techandgraphics.wastical.ui.invoice.pdf.invoiceToPdf
 import net.techandgraphics.wastical.ui.screen.company.client.invoice.CompanyPaymentInvoiceEvent.Button
 import net.techandgraphics.wastical.ui.screen.company.client.invoice.CompanyPaymentInvoiceEvent.Load
 import java.io.File
